fix: move todo tool types to common folder
